cleanup a little
